+++ R/ASCNworkflow.R
@@ -14,7 +14,8 @@
 #' - start: start position of the segment.
 #' - end: end position of the segment.
 #' - total_cn: total integer copy number of the segment.
-#' - minor_cn: minor allele integer copy number of the segment.
+#' - minor_cn: minor allele integer copy number of the segment. Set it
+#' to `NA` if you don't have this data.
 #' - sample: sample identifier.
 #' - purity: tumor purity of the sample. Set to `1` if you don't know.
 #' - ploidy (optinal): ploidy value of the sample tumor genome.
@@ -30,15 +31,20 @@
 #' @examples
 #' data("ascn")
 #' data <- ascn
-#' rv <- gcap.ASCNworkflow(data, outdir = tempdir())
+#' rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
 #' data$purity <- 1
-#' rv2 <- gcap.ASCNworkflow(data, outdir = tempdir())
+#' rv2 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
 #' data$age <- 60
 #' data$gender <- "XY"
-#' rv3 <- gcap.ASCNworkflow(data, outdir = tempdir())
+#' rv3 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB32")
+#' # If you only have total integer copy number
+#' data$minor_cn <- NA
+#' rv4 = gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
 #' @testexamples
 #' expect_equal(rv, rv2)
 #' expect_equal(ncol(rv3$by_gene), 35L)
+#' expect_equal(ncol(rv4$by_gene), 15L)
+#' expect_error(gcap.ASCNworkflow(data, outdir = tempdir()))
 gcap.ASCNworkflow <- function(data,
                               genome_build = c("hg38", "hg19"),
                               model = "XGB32",