+++ R/features.R
@@ -53,16 +53,18 @@ gcap.extractFeatures <- function(ascat_files,
     )
   }
 
+
   lg$info("generating CopyNumber object in sigminer package")
   cn <- data.table::copy(rvlist$data)
   if ("old_sample" %in% colnames(cn)) cn$old_sample <- NULL
+  add_loh = if (all(is.na(cn$minor_cn))) FALSE else TRUE
   cn <- read_copynumber(
     cn,
     seg_cols = colnames(cn)[1:4],
     max_copynumber = 10000L,
     join_adj_seg = FALSE,
     genome_build = genome_build,
-    add_loh = TRUE,
+    add_loh = add_loh,
     loh_min_len = 1e3
   )
 
@@ -82,21 +84,31 @@ gcap.extractFeatures <- function(ascat_files,
     rm_black_arms = TRUE
   )[, .(sample, AScore)]
 
-  lg$info("getting pLOH score")
-  df_pLOH <- get_pLOH_score(cn)
+  if (add_loh) {
+    lg$info("getting pLOH score")
+    df_pLOH <- get_pLOH_score(cn)
+  } else {
+    lg$info("pLOH score is skipped as no LOH data available")
+    df_pLOH = NULL
+  }
 
   lg$info("getting CNA burden")
   df_cna <- cn@summary.per.sample[, .(sample, cna_burden)]
 
-  lg$info("generating copy number catalog matrix for fitting signature activity")
-  cn_tally <- sig_tally(cn, method = "S")
-  lg$info("fitting copy number signature activity")
-  df_act <- sig_fit(
-    t(cn_tally$all_matrices$CN_48),
-    sig_db = "CNS_TCGA",
-    sig_index = "ALL",
-    return_class = "data.table"
-  )
+  if (add_loh) {
+    lg$info("generating copy number catalog matrix for fitting signature activity")
+    cn_tally <- sig_tally(cn, method = "S")
+    lg$info("fitting copy number signature activity")
+    df_act <- sig_fit(
+      t(cn_tally$all_matrices$CN_48),
+      sig_db = "CNS_TCGA",
+      sig_index = "ALL",
+      return_class = "data.table"
+    )
+  } else {
+    lg$info("copy number signature activity quantification is skipped as no LOH data available")
+    df_act = NULL
+  }
 
   lg$info("merging data")
   fts_sample <- mergeDTs(list(